started preparing syncInquirer for refactoring with tests
